+++ packages/social-previews/src/linkedin-preview/constants.ts
@@ -1,0 +1,7 @@
+export const FEED_TEXT_MAX_LENGTH = 212;
+export const FEED_TEXT_MAX_LINES = 3;
+
+export const DEFAULT_PROPS = {
+	name: 'Account Name',
+	profileImage: 'https://static.licdn.com/sc/h/1c5u578iilxfi4m4dvc4q810q',
+};
