+++ packages/social-previews/src/linkedin-preview/types.ts
@@ -1,0 +1,9 @@
+import { SocialPreviewBaseProps, SocialPreviewsBaseProps } from '../types';
+
+export type LinkedInPreviewProps = SocialPreviewBaseProps & {
+	jobTitle?: string;
+	name: string;
+	profileImage: string;
+};
+
+export type LinkedInPreviewsProps = LinkedInPreviewProps & SocialPreviewsBaseProps;
